refactor: consistency accross projects, memoize config, and separation of concerns
