+++ src/main/kotlin/no/nav/emottak/eventmanager/configuration/Configurator.kt
@@ -1,14 +1,17 @@
 package no.nav.emottak.eventmanager.configuration
 
+import arrow.core.memoize
 import com.sksamuel.hoplite.ConfigLoader
 import com.sksamuel.hoplite.addEnvironmentSource
 import com.sksamuel.hoplite.addResourceSource
 
-fun config() = ConfigLoader.builder()
-    .addEnvironmentSource()
-    .addResourceSource("/application-personal.conf", optional = true)
-    .addResourceSource("/kafka_common.conf")
-    .addResourceSource("/application.conf")
-    .withExplicitSealedTypes()
-    .build()
-    .loadConfigOrThrow<Config>()
+val config: () -> Config = {
+    ConfigLoader.builder()
+        .addEnvironmentSource()
+        .addResourceSource("/application-personal.conf", optional = true)
+        .addResourceSource("/kafka_common.conf")
+        .addResourceSource("/application.conf")
+        .withExplicitSealedTypes()
+        .build()
+        .loadConfigOrThrow<Config>()
+}.memoize()
