+++ src/main/kotlin/no/nav/emottak/eventmanager/plugin/Authentication.kt
@@ -1,0 +1,14 @@
+package no.nav.emottak.eventmanager.plugin
+
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.auth.Authentication
+import no.nav.emottak.eventmanager.auth.AZURE_AD_AUTH
+import no.nav.emottak.eventmanager.auth.AuthConfig
+import no.nav.security.token.support.v3.tokenValidationSupport
+
+fun Application.configureAuthentication() {
+    install(Authentication) {
+        tokenValidationSupport(AZURE_AD_AUTH, AuthConfig.getTokenSupportConfig())
+    }
+}
