+++ src/main/kotlin/no/nav/emottak/eventmanager/plugin/ContentNegotiation.kt
@@ -1,0 +1,10 @@
+package no.nav.emottak.eventmanager.plugin
+
+import io.ktor.serialization.kotlinx.json.json
+import io.ktor.server.application.Application
+import io.ktor.server.application.install
+import io.ktor.server.plugins.contentnegotiation.ContentNegotiation
+
+fun Application.configureContentNegotiation() {
+    install(ContentNegotiation) { json() }
+}
