+++ src/main/kotlin/no/nav/emottak/eventmanager/route/NaisRoutes.kt
@@ -1,0 +1,19 @@
+package no.nav.emottak.eventmanager.route
+
+import io.ktor.server.response.respond
+import io.ktor.server.response.respondText
+import io.ktor.server.routing.Routing
+import io.ktor.server.routing.get
+import io.micrometer.prometheus.PrometheusMeterRegistry
+
+fun Routing.naisRoutes(collectorRegistry: PrometheusMeterRegistry) {
+    get("/internal/health/liveness") {
+        call.respondText("I'm alive! :)")
+    }
+    get("/internal/health/readiness") {
+        call.respondText("I'm ready! :)")
+    }
+    get("/internal/prometheus") {
+        call.respond(collectorRegistry.scrape())
+    }
+}
