+++ src/main/resources/application.conf
@@ -1,3 +1,7 @@
+server {
+  port = 8080
+  preWait = "5s"
+}
 
 database {
   vaultJdbcUrl = "${JDBC_URL:-}"
